// ui/src/lib/types/SlideshowConfig.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { SlideshowSort } from './SlideshowSort';
import type { SlideshowStart } from './SlideshowStart';

export type SlideshowConfig = {
  interval_secs: number;
  sort: SlideshowSort;
  recent_history_size: number;
  on_start: SlideshowStart;
  pause_when_active: boolean;
  skip_on_unavailable: boolean;
};