// ui/src/lib/types/SpanSource.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { ImageSet } from './ImageSet';
import type { SlideshowConfig } from './SlideshowConfig';

export type SpanSource =
  | { type: 'single'; path: string }
  | { type: 'slideshow'; images: ImageSet; config: SlideshowConfig };